requirements.txt
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.0
Flask-Login==0.5.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
Jinja2==2.11.2
newsapi-python==0.2.6
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.1